remove parse_date in favour of existing function
